% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCARH.model.R
\name{iCARH.model}
\alias{iCARH.model}
\title{Runs the integrative CAR Horseshoe model}
\usage{
iCARH.model(X, Y, drug, pathways, tau = 1.2, NA_value = -99999, ...)
}
\arguments{
\item{X}{the metabolomics time-course data}

\item{Y}{the additional omic time-course data}

\item{drug}{treatment effect (NA values not allowed in drug)}

\item{pathways}{pathway adjacency matrices}

\item{tau}{global sparsity parameter \eqn{\tau} as in Jendoubi, T., & Ebbels, T. (2018)}

\item{NA_value}{NA values are incompatible with stan.
NAs will be replaced by NA_value and will be inferred (only for X and Y data).}

\item{...}{additional stan parameters}
}
\value{
stan object
}
\description{
Infers treatment effects, association with heterogenous omic variables, pathway perturbation
among other parameters (e.g. time dependance). Regression coefficients (beta parameter) are initialised
using a univariate regression ignoring time and metabolite dependence.
}
\examples{
data.sim = iCARH.simulate(4, 8, 10, 2, 2, path.probs=0.3, Zgroupeff=c(0,4),
beta.val=c(1,-1,0.5, -0.5))
XX = data.sim$XX
Y = data.sim$Y
Z = data.sim$Z
pathways = data.sim$pathways
fit = iCARH.model(XX, Y, Z, pathways, control = list(adapt_delta = 0.99, max_treedepth=10),
iter = 500, chains = 2)

}
